+++ tests/expresso_test.py
@@ -57,10 +57,19 @@ def test_lists():
     assert node is not None
 
     expr = "searchsorted(['9E', 'AA', 'AQ', 'AS', 'B6', 'CO', 'DL', 'EV', 'F9', 'FL', 'HA', 'MQ', 'NW', 'OH', 'OO', 'UA', 'US', 'WN', 'XE', 'YV'], UniqueCarrier)"
+    validate_expression(expr, {'UniqueCarrier'}, {'searchsorted'})
     expr_translate = translate(expr, lambda x: None)
     print(node)
     assert expr == expr_translate
 
+def test_dicts():
+    expr = "fillmissing(o, {'a': 1})"
+    validate_expression(expr, {'o'}, {'fillmissing'})
+    node = parse_expression(expr)
+    assert node is not None
+    expr_translate = translate(expr, lambda x: None)
+    print(node)
+    assert expr == expr_translate
 
 def test_validate():
     validate_expression('x + 1', {'x'})